ver 0.9.9 (2009-11-23) 1. 调整：多语言内核合并成一个发布包，同时提供：简体、繁体和英文，并对内核进行了针对性修改 2. 添加：工具栏改为自动换行模式，添加工具栏强制换行命令：BtnBr，并对相关的样式和脚本进行了针对性修改 3. 调整：自带Ajax上传开始后弹出假上传进度窗口，防止用户意外关闭上传前操作界面导致上传失败 4. 添加：增加缩进为Tab快捷键，减少缩进为Shift+Tab快捷键 5. 修正：编辑内容区在预览状态下点击链接，会导致当前编辑区域失效，现更改为默认打开新窗口(_target) 6. 修正：演示用upload.php增加扩展名限制，以增加文件上传的安全性 7. 添加：UBB模式增加[code]的完美支持，保证[code][/code]区域里的内容不做任何修改 8. 修正：图片属性界面中的水平间距和垂直间距反了，在此重新更正 9. 添加：在settings中添加baseUrl参数，应用在需要相对路径的编辑环境中：编辑页面和图片文件不在同一个基础路径下，但是又需要图片等附件保存为相对路径 10.修正：Firefox下切换到源代码再切换回来，会导致无法用删除键和后退键问题的修正 11.调整：showModal函数由原先的单一iframe上传功能，现分离成两个独立函数：showIframeModal和showModal，前者调用内嵌iframe的模式窗口，后者显示普通的模式窗口，都可供插件调用 12.添加：添加getSelect函数，返回编辑器窗口中选中内容的HTML代码或者文本，唯一参数设置为'text'即是返回文本 13.添加：添加link链接的附件模式，不选择任何内容的情况下添加链接为添加附件，附件链接的文字可设置参数attachLinkText 14.修正：强制P标签功能的进一步完善，非P标签模式下，按回车键全部输出换行 15.添加：添加toggleShowBlocktag接口，以方便某些需要添加“显示或者隐藏块标签”按钮的开发者调用 16.修正：formatXHTML兼容性进一步提高，处理某些非标准HTML代码，导致部分内容丢失问题的修正 17.添加：系统自带的Ajax上传和iframe式上传增加“立即添加模式”，只需在返回的参数最前面加感叹号“!”，在上传完成后图片或者链接等内容即自动加入到编辑器内容中 18.修正：系统自带Ajax上传功能在Opera上传失败问题的修正

+++ trunk/demos/upload.php
@@ -16,6 +16,7 @@ function uploadfile($inputname)
 	$attachdir='upload';//上传文件保存路径，结尾不要带/
 	$dirtype=1;//1:按天存入目录 2:按月存入目录 3:按扩展名存目录  建议使用按天存
 	$maxattachsize=2097152;//最大上传大小，默认是2M
+	$upext='txt,rar,zip,jpg,jpeg,gif,png,swf,avi';//上传扩展名
 	
 	$err = "";
 	$msg = "";
@@ -50,38 +51,39 @@ function uploadfile($inputname)
 				$err = '无有效错误代码';
 		}
 	}
-	elseif(empty($upfile['tmp_name']) || $upfile['tmp_name'] == 'none')
-	{
-		$err = '无文件上传';
-	}
+	elseif(empty($upfile['tmp_name']) || $upfile['tmp_name'] == 'none')$err = '无文件上传';
 	else
 	{
 			$temppath=$upfile['tmp_name'];
 			$fileinfo=pathinfo($upfile['name']);
 			$extension=$fileinfo['extension'];
-			$filesize=filesize($temppath);
-			if($filesize <= $maxattachsize)
+			if(preg_match('/'.str_replace(',','|',$upext).'/i',$extension))
 			{
-				switch($dirtype)
+				$filesize=filesize($temppath);
+				if($filesize <= $maxattachsize)
 				{
-					case 1: $attach_subdir = 'day_'.date('ymd'); break;
-					case 2: $attach_subdir = 'month_'.date('ym'); break;
-					case 3: $attach_subdir = 'ext_'.$extension; break;
+					switch($dirtype)
+					{
+						case 1: $attach_subdir = 'day_'.date('ymd'); break;
+						case 2: $attach_subdir = 'month_'.date('ym'); break;
+						case 3: $attach_subdir = 'ext_'.$extension; break;
+					}
+					$attach_dir = $attachdir.'/'.$attach_subdir;
+					if(!is_dir($attach_dir))
+					{
+						@mkdir($attach_dir, 0777);
+						@fclose(fopen($attach_dir.'/index.htm', 'w'));
+					}
+					PHP_VERSION < '4.2.0' && mt_srand((double)microtime() * 1000000);
+					$filename=date("YmdHis").mt_rand(1000,9999).'.'.$extension;
+					$target = $attach_dir.'/'.$filename;
+					
+					move_uploaded_file($upfile['tmp_name'],$target);
+					$msg=$target;
 				}
-				$attach_dir = $attachdir.'/'.$attach_subdir;
-				if(!is_dir($attach_dir))
-				{
-					@mkdir($attach_dir, 0777);
-					@fclose(fopen($attach_dir.'/index.htm', 'w'));
-				}
-				PHP_VERSION < '4.2.0' && mt_srand((double)microtime() * 1000000);
-				$filename=date("YmdHis").mt_rand(1000,9999).'.'.$extension;
-				$target = $attach_dir.'/'.$filename;
-				
-				move_uploaded_file($upfile['tmp_name'],$target);
-				$msg=$target;
+				else $err='文件大小超过'.$maxattachsize.'字节';
 			}
-			else $err='文件大小超过'.$maxattachsize.'字节';
+			else $err='上传文件扩展名必需为：'.$upext;
 
 			@unlink($temppath);
 	}
@@ -90,6 +92,5 @@ function uploadfile($inputname)
 
 $state=uploadfile('upload');
 echo json_encode($state);
-
 
 ?>
